Fix deadlock in quotes
