+++ lib/luro_commands/src/commands/quote.rs
@@ -1,45 +1,35 @@
-use luro_core::quotes::{Quote, Quotes};
-use luro_core::{Context, Error, QUOTES_FILE_PATH};
-use luro_utilities::guild_accent_colour;
+use luro_core::quotes::Quote;
+use luro_core::{Context, Error};
 use poise::serenity_prelude::Message;
 
 use crate::commands::quote::get::get;
 use crate::commands::quote::user::user;
+use crate::commands::quote::send_quote::send_quote;
+use crate::commands::quote::save_quote::save_quote;
 
 mod get;
 mod user;
+mod save_quote;
+mod send_quote;
 
 /// Get some information on things, like guilds and users.
 #[poise::command(
     context_menu_command = "Save this quote",
     slash_command,
-    category = "Guild",
+    category = "Quotes",
     subcommands("get", "user")
 )]
 pub async fn quote(
     ctx: Context<'_>,
     #[description = "The quote which you wish to add to the database"] message: Message
 ) -> Result<(), Error> {
-    let accent_colour = ctx.data().config.read().await.accent_colour;
-    let quotes = &ctx.data().quotes.read().await.quotes;
-    let mut new_quote = vec![Quote {
+    let new_quote = vec![Quote {
         user_id: *message.author.id.as_u64(),
         quote: String::from(&message.content)
     }];
 
-    ctx.send(|b| {
-        b.embed(|b| {
-            b.author(|a| a.name(&message.author.name).icon_url(&message.author.face()))
-                .title("Quote Added!")
-                .description(&message.content)
-                .color(guild_accent_colour(accent_colour, ctx.guild()))
-                .footer(|f| f.text(format!("Quote ID: {}", quotes.len())))
-        })
-    })
-    .await?;
-    let quotes = &mut ctx.data().quotes.write().await;
-    quotes.quotes.append(&mut new_quote);
-    Quotes::write(quotes, QUOTES_FILE_PATH).await;
+    send_quote(ctx, message, "Quote Added!".into(), None).await?;
+    save_quote(ctx.data().quotes.clone(), new_quote).await;
 
     Ok(())
 }
