+++ lib/luro_commands/src/commands/quote/get.rs
@@ -10,23 +10,18 @@ pub async fn get(ctx: Context<'_>, #[description = "Get a quote by ID"] quote: O
     let quotes = &ctx.data().quotes.read().await.quotes;
     let random_number = rand::thread_rng().gen_range(0..quotes.len());
 
-    // Try to get the specified quote
-    if let Some(quote) = quote {
-        if let Some(quote_resolved) = quotes.get(quote) {
-            send_quote(ctx, quote_resolved, Some(quote)).await?;
-            return Ok(());
-        } else {
+    let returned_quote = match quote {
+        Some(quote_defined) => quotes.get(quote_defined),
+        None => quotes.get(random_number),
+    };
+
+    match returned_quote {
+        Some(quote_resolved) => send_quote(ctx, quote_resolved, quote).await?,
+        None => {
             ctx.say("Failed to get that quote! Sure you got the right ID?").await?;
             return Ok(());
-        }
-    }
-
-    // No quote specified, so lets get a random one instead!
-    if let Some(quote) = quotes.get(random_number) {
-        send_quote(ctx, quote, Some(random_number)).await?;
-    } else {
-        ctx.say("Failed to find a quote").await?;
-    }
+        },
+    };
 
     Ok(())
 }
